requirements.txt
protobuf
google-auth
google-auth-oauthlib
Flask
markdown
openai
flask[async]